utils/requirements.txt
nibabel~=5.1.0
tqdm~=4.65.0
numpy~=1.24.3
scikit-learn~=1.3.0
torchvision~=0.16.1